requirements.txt
feedparser
pandas
nltk
uvicorn[standard]
scikit-learn
joblib
fastapi
pydantic
transformers
torch
requests
python-dotenv
openai
faiss-cpu
sentence_transformers
python-multipart
